Rename pkg and auto create dirs
