--- src/main/java/org/fairgenomes/generator/implementations/rdfowl/ToTTL.java
@@ -1,4 +1,0 @@
-package org.fairgenomes.generator.implementations.rdfowl;
-
-public class ToTTL {
-}
