+++ src/main/java/org/fairgenomes/transformer/Main.java
@@ -1,11 +1,11 @@
-package org.fairgenomes.generator;
+package org.fairgenomes.transformer;
 
 public class Main {
 
     public static void main(String args[]) throws Exception {
         System.out.println("Starting...");
         long start = System.nanoTime();
-        Generator s = new Generator();
+        Transformer s = new Transformer();
         s.generateResources();
         System.out.println("Done! Completed in " + ((System.nanoTime()-start)/1000000)+"ms.");
     }
