+++ src/main/java/org/fairgenomes/transformer/datastructures/Element.java
@@ -1,4 +1,4 @@
-package org.fairgenomes.generator.datastructures;
+package org.fairgenomes.transformer.datastructures;
 
 public class Element {
