+++ src/main/java/org/fairgenomes/transformer/datastructures/ValueType.java
@@ -1,4 +1,4 @@
-package org.fairgenomes.generator.datastructures;
+package org.fairgenomes.transformer.datastructures;
 
 public enum ValueType {
     String, LookupOne, LookupMultiple, PositiveInteger, CrossReference, Date, DateTime, Boolean, Float;
