+++ src/main/java/org/fairgenomes/transformer/implementations/rdfowl/ToTTL.java
@@ -1,0 +1,4 @@
+package org.fairgenomes.transformer.implementations.rdfowl;
+
+public class ToTTL {
+}
